Refactor dockerfile and install front end app by default

+++ scripts/docker-entrypoint.sh
@@ -1,17 +1,26 @@
 #!/bin/bash
-# CMD to be run inside the g3w-suite docker, this script is
-# volume-mounted (ro) in the docker compose as /cmd.sh
+# Entrypoint script
 
 # Start XVfb
-rm /tmp/.X99-lock
+if [[  -f /tmp/.X99-lock ]]; then
+  rm /tmp/.X99-lock
+fi
 Xvfb :99 -screen 0 640x480x24 -nolisten tcp &
 export DISPLAY=:99
 export QGIS_SERVER_PARALLEL_RENDERING=1
+# Start
+cd /code/g3w-admin
+
+rm -rf /shared-volume/build_done
+
 # Build the suite
 /code/ci_scripts/build_suite.sh
 # Setup once
 /code/ci_scripts/setup_suite.sh
-# Start
-cd /code/g3w-admin
+# Run migrations to activate the front end app
+python3 manage.py makemigrations
+python3 manage.py migrate
+python3 manage.py collectstatic --noinput
+
 gunicorn base.wsgi:application --limit-request-fields 0 --error-logfile - \
     --log-level=debug --timeout 120 --workers=${G3WSUITE_GUNICORN_NUM_WORKERS:-8} -b 0.0.0.0:8000
